+++ icarving.api.pinche/src/main/java/cn/icarving/api/pinche/dto/PickActivityForm.java
@@ -5,7 +5,7 @@ import java.sql.Timestamp;
 
 public class PickActivityForm {
 
-	private long ownerId;
+	private int ownerId;
 
 	private Timestamp startTime;
 
@@ -23,11 +23,11 @@ public class PickActivityForm {
 
 	private String note;
 
-	public long getOwnerId() {
+	public int getOwnerId() {
 		return ownerId;
 	}
 
-	public void setOwnerId(long ownerId) {
+	public void setOwnerId(int ownerId) {
 		this.ownerId = ownerId;
 	}
